+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 - Enh #62: Add method `StringHelper::split` that split a string to array with non-empty lines (vjik)
 - Enh #64: Add support of a boolean values to `NumericHelper::normalize` (vjik) 
+- Enh #63: Add method `NumericHelper::isInteger` that checks whether the given string is an integer number (vjik)
 
 ## 1.1.0 November 13, 2020
 
@@ -20,6 +21,5 @@
 ## 1.0.0 August 31, 2020
 
 - Initial release.
-