+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Yii Strings Change Log
 
 
+## 1.2.1 under development
+
+- no changes in this release.
+
+
 ## 1.2.0 January 22, 2021
 
 - Enh #62: Add method `StringHelper::split()` that split a string to array with non-empty lines (vjik)